Simplify base `changeListener`.
When handling resync of stale cursors with the underlying immutable structure (e.g map),
we need to handle three cases based on the given keyPath indicating a change:

1. keyPath is not in the updated map, but it exists in the current (removal)
2. keyPath is in the updated map, but it doesn't exist in the current (add)
3. keyPath is in the updated map, but it also exists in the current

For the first case, we need to use `Immutable.Map.removeIn`.
While `Immutable.Map.setIn` is sufficient for the other cases.

Using methods of simplifying boolean expressions (e.g. K-maps), we can just test
keyPath is not in the updated map to handle logic flow.